+++ CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Testing for single script file (Add `execute` evaluation).
+
+### Removed
+
+- Drop support for busybox 1.1.3 and similar versions as it can not redefine builtin commands.
+
+### Added
+
 - Add installer
 
 ## [0.18.0] - 2019-07-09