+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.19.0] - 2019-07-20
+
 ### Added
 
 - Add installer (It has not been officially released, but you can used it already).
@@ -218,7 +220,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial public release.
 
-[Unreleased]: https://github.com/shellspec/shellspec/compare/0.18.0...HEAD
+[Unreleased]: https://github.com/shellspec/shellspec/compare/0.19.0...HEAD
+[0.19.0]: https://github.com/shellspec/shellspec/compare/0.18.0...0.19.0
 [0.18.0]: https://github.com/shellspec/shellspec/compare/0.17.0...0.18.0
 [0.17.0]: https://github.com/shellspec/shellspec/compare/0.16.0...0.17.0
 [0.16.0]: https://github.com/shellspec/shellspec/compare/0.15.0...0.16.0